[APP-2479] Move some components to the design system package
